Fixed the monad dependency issue in RetryT.
Need to add 'm' as an extra type parameter due to the dependence of
RetryT on the monad's type when executing 'withRetry' on a block. For
more details, see here:

https://www.reddit.com/r/haskell/comments/3w7vze/help_with_freet_transformer/

Thanks to Edward Kmett for the help!

However, note that the 'WithRetry' command as currently implemented in
'runRetryT' does not actually retry! It will simply fail, because
nothing is there to catch the failure. This is because, as Edward points
out in the cited Reddit discussion, it may not be safe to retry the
block due to side effects in the monad which cannot be un-done upon
retry. Therefore, semantically, we can say that (RetryT m) over MonadIO
instances may not make sense. It would be interesting to re-target
runRetryT to the STM monad, since in that monad we can back out of
transactions upon failure. However, for the purposes of this exercise,
which was simply to get a FreeT transformer example working with a
'with'-style command, the problem is now solved, and I'm going to commit
this fix as-is.